+++ hardware.h
@@ -10,8 +10,8 @@ int doit_attendre();*/
 void init_hardware();
 void set_PWM_moteur_D(int PWM);
 void set_PWM_moteur_G(int PWM);
-int get_nbr_tick_D();
-int get_nbr_tick_G();
+long int get_nbr_tick_D();
+long int get_nbr_tick_G();
 int doit_attendre();
 
 #endif
